premiere_league_data_api/requirements.txt
requests
pandas
matplotlib
openpyxl
python-dotenv